requirements.txt
alembic==1.0.11
astroid==2.2.5
autopep8==1.4.4
bcrypt==3.1.6
blinker==1.4
cffi==1.12.2
Click==7.0
entrypoints==0.3
flake8==3.7.7
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
future==0.17.1
gunicorn==19.9.0
iso8601==0.1.12
isort==4.3.16
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.3.1
Mako==1.0.12
MarkupSafe==1.1.1
mccabe==0.6.1
Pillow==6.0.0
pkg-resources==0.0.0
psycopg2-binary==2.8.3
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
pylint==2.3.1
pylint-flask==0.6
pylint-plugin-utils==0.5
pyserial==3.4
python-dateutil==2.8.0
python-dotenv==0.10.3
python-editor==1.0.4
PyYAML==5.1
six==1.12.0
SQLAlchemy==1.3.2
typed-ast==1.3.1
Werkzeug==0.15.2
wrapt==1.11.1
WTForms==2.2.1
